Revert "zsk/extract gencase"
